(Slight) rework for damage detection for Tank selection
Player selection for Tank spawns has always worked as intended - it chooses the player who has dealt the most damage for that round - but the way it sees how much damage a player has dealt can be questionable at times.

How it works now: Everyone (obviously) starts at 0 damage and all damage dealt is added towards a damage counter, even damage done as a survivor. Players who joined the server mid-round have their damage counter set to the average damage dealt by all players, **including survivors**. When a Tank dies, that player's damage counter is also set to the average damage dealt by all players.

This means that players who died or (especially) joined the server mid-round have a much greater chance to become a Tank, as damage done by survivors greatly exceed damage done by infected. It's also common to see a player become a Tank multiple times in a row because of this.

This PR does a couple of changes:
- Removed the 'average damage' function altogether, setting damage counters to 0 in these cases instead.
- Made Tank selection only consider infected performance.

This does add a little incentive for players who start as infected to deal damage as they get this head start for Tank consideration, and the more incentive to get staring infected to actually play the game, the better.