Hunter: Longer Pounce Cooldown
Increases rage cooldown by three seconds (aka substantially increasing cooldown on missing a pounce). Probably one of the best special infected currently in the mode for being able to lock down and potentially kill a survivor by themselves. Only having a three second cooldown on missing a pounce is too generous, especially compared to Jockey which has a longer cooldown on miss with a similar, but more situational power.